backend-library/library/requirements.txt
anyio==4.5.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
cryptography==43.0.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.15.4
fqdn==1.5.1
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.9
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netmiko==4.4.0
notebook==7.2.2
notebook_shim==0.2.4
ntc_templates==6.0.0
overrides==7.7.0
packaging==24.1
pandocfilters==1.5.1
paramiko==3.4.1
parso==0.8.4
pipenv==2024.0.1
platformdirs==4.2.2
ply==3.11
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyasn1==0.6.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
PyNaCl==1.5.0
pyserial==3.5
pysmi==1.5.0
pysnmp==7.1.2
pysnmpcrypto==0.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
scp==0.15.0
Send2Trash==1.8.3
setuptools==71.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
terminado==0.18.1
textfsm==1.1.3
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20240906
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.3
virtualenv==20.26.3
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
